+++ RELEASE_NOTES.md
@@ -16,6 +16,7 @@
 * Added prebuilt bindings for `Selector.SelectedIndex`
 * Added Binding.TwoWay.id to API
 * Removed trace logging during a successful set in a `SubModelSelectedItem` binding
+* Added one-way-to-source binding
 
 #### 4.0.0-beta-41
 * Fixed "backwards typing" (#373) and other bugs (like #371) introduced in 4.0.0-beta-40 with more careful use of the Dispatcher (#374)